Added Joi validation

+++ backend/routes/tasksRouter.js
@@ -5,11 +5,16 @@ const { isAuthenticated } = require("../auth/jwt-auth");
 const { createDocument, getDocumentWithId, createDocumentWithId, updateDocument, deleteDocument } = require('../controllers/documentController');
 const { generateFuzzyArray } = require('../controllers/fuzzyController');
 const { getMediaURL, deleteMediaObject } = require('../controllers/storageController');
-const { or, query, where, collection, getDocs, and } = require('firebase/firestore');
+const { or, query, where, collection, getDocs, and, orderBy } = require('firebase/firestore');
+const { createTaskSchema, getTasksSchema, updateTaskSchema, deleteTaskSchema } = require('../schemas/taskSchemas');
 
 // Tasks CRUD
 router.post("/", isAuthenticated, async (req, res) => {
     try {
+        const { error, value } = createTaskSchema.validate(req.body);
+        if (error) {
+            return res.status(400).json({ message: error.details[0].message });
+        }
         const { title, description, dueDate, priority, attachments } = req.body
 
         const titleFuzzy = generateFuzzyArray(title);
@@ -21,7 +26,7 @@ router.post("/", isAuthenticated, async (req, res) => {
             titleFuzzy,
             descriptionFuzzy,
             dueDate: new Date(dueDate),
-            priority,
+            priority: priorityMap({priorityString: priority}),
             attachments,
             status: "New",
             user: req.userRef
@@ -34,12 +39,19 @@ router.post("/", isAuthenticated, async (req, res) => {
 
 router.get("/", isAuthenticated, async (req, res) => {
     try {
-        const { searchTerm, status, id } = req.query
+        const { error, value } = getTasksSchema.validate(req.query);
+        if (error) {
+            return res.status(400).json({ message: error.details[0].message });
+        }
+        const { searchTerm, status, priority, id } = req.query
         const data = [];
         let queries = [where("user", "==", req.userRef)];
 
         if (id) {
             const task = await getDocumentWithId("tasks", id);
+            if(!task.exists){
+                throw Error("Task not found")
+            }
             const taskAttachments = await Promise.all(task.attachments.map(async (attachment) => {
                 return await getMediaURL(attachment);
             }))
@@ -54,11 +66,15 @@ router.get("/", isAuthenticated, async (req, res) => {
             queries.push(where("status", "==", status));
         }
 
-        const querySnapshot = await getDocs(query(collection(firestore, "tasks"), and(...queries)));
+        if (priority) {
+            queries.push(where("priority", "==", priorityMap({priorityString: priority})));
+        }
+
+        const querySnapshot = await getDocs(query(collection(firestore, "tasks"), and(...queries), orderBy("priority", "desc")));
 
         querySnapshot.forEach(async (doc) => {
             const docData = doc.data()
-            data.push(projectTask({ task: {id: doc.id, ...docData} }))
+            data.push(projectTask({ task: { id: doc.id, ...docData } }))
         })
 
         return res.status(200).json({ tasks: data });
@@ -69,9 +85,15 @@ router.get("/", isAuthenticated, async (req, res) => {
 
 router.put("/", isAuthenticated, async (req, res) => {
     try {
+        const { error, value } = updateTaskSchema.validate(req.body);
+        if (error) {
+            return res.status(400).json({ message: error.details[0].message });
+        }
         const { id, title, description, status, priority, attachments } = req.body
-
         const task = await getDocumentWithId("tasks", id);
+        if(!task.exists){
+            throw Error("Task not found")
+        }
         const deletedAttachments = task.attachments.filter((attachment) => !attachments.includes(attachment))
         await Promise.all(deletedAttachments.map(async (attachment) => { return await deleteMediaObject(attachment) }))
 
@@ -79,7 +101,7 @@ router.put("/", isAuthenticated, async (req, res) => {
             ...(title ? { title, titleFuzzy: generateFuzzyArray(title) } : {}),
             ...(description ? { description, descriptionFuzzy: generateFuzzyArray(description) } : {}),
             ...(status ? { status } : {}),
-            ...(priority ? { priority } : {}),
+            ...(priority ? { priority: priorityMap({priorityString: priority}) } : {}),
             ...(attachments ? { attachments } : {}),
         }, id);
 
@@ -95,9 +117,15 @@ router.put("/", isAuthenticated, async (req, res) => {
 
 router.delete("/", isAuthenticated, async (req, res) => {
     try {
+        const { error, value } = deleteTaskSchema.validate(req.query);
+        if (error) {
+            return res.status(400).json({ message: error.details[0].message });
+        }
         const { id } = req.query
-
         const task = await getDocumentWithId("tasks", id);
+        if(!task.exists){
+            throw Error("Task not found")
+        }
         await Promise.all(task.attachments.map(async (attachment) => { return await deleteMediaObject(attachment) }))
 
         await deleteDocument("tasks", id);
@@ -109,13 +137,29 @@ router.delete("/", isAuthenticated, async (req, res) => {
 })
 
 //Functions
+function priorityMap({ priorityNum, priorityString }) {
+    if(priorityNum) {
+        switch(priorityNum){
+            case 1: return "Low";
+            case 2: return "Medium";
+            case 3: return "High";
+        }
+    } else {
+        switch(priorityString){
+            case "Low": return 1;
+            case "Medium": return 2;
+            case "High": return 3;
+        }
+    }
+}
+
 function projectTask({ task, taskAttachments }) {
     return {
         id: task.id,
         title: task.title,
         description: task.description,
         dueDate: new Date(task.dueDate.seconds * 1000),
-        priority: task.priority,
+        priority: priorityMap({priorityNum: task.priority}),
         status: task.status,
         ...(taskAttachments ? { attachments: taskAttachments } : {}),
         createdAt: new Date(task.createdAt.seconds * 1000),
